Move lettuce under the 'redis' package
