+++ agent/agent-plugins/redis-lettuce-5.x/src/main/java/org/bithon/agent/plugin/redis/lettuce/interceptor/RedisAsyncCommandComplete.java
@@ -14,14 +14,14 @@
  *    limitations under the License.
  */
 
-package org.bithon.agent.plugin.lettuce.interceptor;
+package org.bithon.agent.plugin.redis.lettuce.interceptor;
 
 import io.lettuce.core.protocol.AsyncCommand;
 import org.bithon.agent.instrumentation.aop.IBithonObject;
 import org.bithon.agent.instrumentation.aop.context.AopContext;
 import org.bithon.agent.instrumentation.aop.interceptor.declaration.AfterInterceptor;
 import org.bithon.agent.observability.metric.domain.redis.RedisMetricRegistry;
-import org.bithon.agent.plugin.lettuce.LettuceAsyncContext;
+import org.bithon.agent.plugin.redis.lettuce.LettuceAsyncContext;
 
 /**
  * @author frankchen
