+++ agent/agent-plugins/redis-lettuce-5.x/src/main/java/org/bithon/agent/plugin/redis/lettuce/interceptor/RedisClientConnect.java
@@ -14,36 +14,36 @@
  *    limitations under the License.
  */
 
-package org.bithon.agent.plugin.lettuce.interceptor;
+package org.bithon.agent.plugin.redis.lettuce.interceptor;
 
-import io.lettuce.core.api.StatefulConnection;
+import io.lettuce.core.RedisURI;
 import org.bithon.agent.instrumentation.aop.IBithonObject;
 import org.bithon.agent.instrumentation.aop.context.AopContext;
 import org.bithon.agent.instrumentation.aop.interceptor.declaration.AfterInterceptor;
-import org.bithon.agent.plugin.lettuce.LettuceAsyncContext;
+import org.bithon.agent.observability.utils.HostAndPort;
 import org.bithon.component.commons.utils.ReflectionUtils;
+
 
 /**
  * @author frankchen
  */
-public class RedisAsyncCommandDispatch extends AfterInterceptor {
+public class RedisClientConnect extends AfterInterceptor {
 
     @Override
     public void after(AopContext aopContext) {
-        if (!(aopContext.getReturning() instanceof IBithonObject)) {
+        if (aopContext.hasException()) {
             return;
         }
-        IBithonObject result = (IBithonObject) aopContext.getReturning();
 
-        LettuceAsyncContext asyncContext = new LettuceAsyncContext();
-        asyncContext.setStartTime(System.nanoTime());
-        result.setInjectedObject(asyncContext);
-
-        StatefulConnection<?, ?> connection = ((StatefulConnection<?, ?>) ReflectionUtils.getFieldValue(aopContext.getTarget(),
-                                                                                                        "connection"));
+        Object connection = aopContext.getReturning();
         if (connection instanceof IBithonObject) {
-            String endpoint = (String) ((IBithonObject) connection).getInjectedObject();
-            asyncContext.setEndpoint(endpoint);
+            // forward endpoint to connection
+            RedisURI uri = ((RedisURI) ReflectionUtils.getFieldValue(aopContext.getTarget(), "redisURI"));
+
+            // Since RedisClient allows passing RedisURI to connect method
+            // it's a little bit complex to intercept this method to keep HostAndPort on RedisClient
+            // So, we always construct a HostAndPort string here
+            ((IBithonObject) connection).setInjectedObject(HostAndPort.of(uri.getHost(), uri.getPort()));
         }
     }
 }
